+++ app/src/main/java/com/example/assignmenttest/data/repository/AuthRepository.kt
@@ -1,4 +1,4 @@
-package com.example.assignmenttest
+package com.example.assignmenttest.data.repository
 
 
 interface AuthRepository {
