+++ app/src/main/java/com/example/assignmenttest/data/repository/AuthRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.example.assignmenttest
+package com.example.assignmenttest.data.repository
 
 import com.google.firebase.auth.FirebaseAuth
 import kotlinx.coroutines.tasks.await
@@ -6,7 +6,7 @@ import javax.inject.Inject
 
 class AuthRepositoryImpl @Inject constructor(
     private val firebaseAuth : FirebaseAuth
-): AuthRepository{
+): AuthRepository {
 
     // 회원가입
     override suspend fun signUp(email: String, password: String, nickName: String): Result<Unit> {
