+++ app/src/main/java/com/example/assignmenttest/di/HiltApplication.kt
@@ -1,4 +1,4 @@
-package com.example.assignmenttest
+package com.example.assignmenttest.di
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
