+++ app/src/main/java/com/example/assignmenttest/presentation/fragment/CompleteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.assignmenttest
+package com.example.assignmenttest.presentation.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,8 +8,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
+import com.example.assignmenttest.AuthUiState
+import com.example.assignmenttest.R
 import com.example.assignmenttest.databinding.FragmentCompleteBinding
-import com.example.assignmenttest.databinding.FragmentLoginBinding
+import com.example.assignmenttest.presentation.viewModel.AuthViewModel
 import com.google.android.material.snackbar.Snackbar
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
