+++ app/src/main/res/layout/fragment_complete.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main_yellow"
-    tools:context=".CompleteFragment">
+    tools:context=".presentation.fragment.CompleteFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gl_start"
