+++ app/src/main/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/background_gray"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".LoginFragment">
+    tools:context=".presentation.fragment.LoginFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gl_top"
