+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main_yellow"
-    tools:context=".MainFragment">
+    tools:context=".presentation.fragment.MainFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gl_start"
