+++ app/src/main/res/layout/fragment_signup.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_gray"
-    tools:context=".SignUpFragment">
+    tools:context=".presentation.fragment.SignUpFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/gl_top"
